source/examples/pytest_demo/requirement_dev.txt

coverage==4.5.3
pytest==4.1.1
pytest-cov==2.7.1
pytest-sugar==0.9.2
pytest-tornado==0.8.0
pytest-html==1.20.0
